requirements.txt
fastapi==0.108.0
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.32.0.post1
uvicorn==0.25.0
aiosmtplib==3.0.1